cart updated

+++ index.js
@@ -22,16 +22,16 @@ console.log(quantityOne);
 
 var plusButtonOne = document.getElementsByClassName('plus-one')[0];
 
-var minusButtonOne = document.getElementsByClassName('minus-one')[0];    
+var minusButtonOne = document.getElementsByClassName('minus-one')[0];
 // declaring the action of the plusOne button
 plusButtonOne.addEventListener('click', function () {
-        console.log("Has sumado");
+    console.log("Has sumado");
 
-        quantityOne = quantityOne + 1;
-        console.log(quantityOne);
+    quantityOne = quantityOne + 1;
+    console.log(quantityOne);
 
-        document.getElementsByClassName('item-quantity-one')[0].innerHTML = quantityOne;
-        updateCartTotal ()
+    document.getElementsByClassName('item-quantity-one')[0].innerHTML = quantityOne;
+    updateCartTotal()
 
 })
 
@@ -45,14 +45,14 @@ minusButtonOne.addEventListener('click', function () {
         console.log(quantityOne);
 
         document.getElementsByClassName('item-quantity-one')[0].innerHTML = quantityOne;
-        updateCartTotal ()
-    }else {
+        updateCartTotal()
+    } else {
         quantityOne = 1;
-        updateCartTotal ()
+        updateCartTotal()
     }
-    
 
-}) 
+
+})
 
 
 // Item quantityTwo  ------> Adding stock to the 2nd item
@@ -63,19 +63,19 @@ var quantityTwo = parseInt(getItemQuantityTwo.innerHTML);
 
 console.log(quantityTwo);
 
-var plusButtonTwo = document.getElementsByClassName('plus-two')[0];    
-var minusButtonTwo = document.getElementsByClassName('minus-two')[0];    
+var plusButtonTwo = document.getElementsByClassName('plus-two')[0];
+var minusButtonTwo = document.getElementsByClassName('minus-two')[0];
 // declaring the action of the plus button
 plusButtonTwo.addEventListener('click', function () {
-        console.log("Has sumado");
+    console.log("Has sumado");
 
-        quantityTwo = quantityTwo + 1;
-        console.log(quantityTwo);
+    quantityTwo = quantityTwo + 1;
+    console.log(quantityTwo);
 
-        document.getElementsByClassName('item-quantity-two')[0].innerHTML = quantityTwo;
-        updateCartTotal ()
+    document.getElementsByClassName('item-quantity-two')[0].innerHTML = quantityTwo;
+    updateCartTotal()
 
-})   
+})
 
 // declaring the action of the minusTwo button
 minusButtonTwo.addEventListener('click', function () {
@@ -87,14 +87,14 @@ minusButtonTwo.addEventListener('click', function () {
         console.log(quantityTwo);
 
         document.getElementsByClassName('item-quantity-two')[0].innerHTML = quantityTwo;
-        updateCartTotal ()
-    }else {
+        updateCartTotal()
+    } else {
         quantityTwo = 1;
-        updateCartTotal ()
+        updateCartTotal()
     }
-    
 
-}) 
+
+})
 
 
 
@@ -109,20 +109,20 @@ var quantityThree = parseInt(getItemQuantityThree.innerHTML);
 
 console.log(quantityThree);
 
-var plusButtonThree = document.getElementsByClassName('plus-three')[0];    
-var minusButtonThree = document.getElementsByClassName('minus-three')[0];    
+var plusButtonThree = document.getElementsByClassName('plus-three')[0];
+var minusButtonThree = document.getElementsByClassName('minus-three')[0];
 // declaring the action of the plus button
 plusButtonThree.addEventListener('click', function () {
-        console.log("Has sumado");
+    console.log("Has sumado");
 
-        quantityThree = quantityThree + 1;
-        console.log(quantityThree);
+    quantityThree = quantityThree + 1;
+    console.log(quantityThree);
 
-        document.getElementsByClassName('item-quantity-three')[0].innerHTML = quantityThree;
+    document.getElementsByClassName('item-quantity-three')[0].innerHTML = quantityThree;
 
-        updateCartTotal ()
+    updateCartTotal()
 
-})   
+})
 // declaring the action of the minusThree button
 minusButtonThree.addEventListener('click', function () {
 
@@ -134,14 +134,14 @@ minusButtonThree.addEventListener('click', function () {
 
         document.getElementsByClassName('item-quantity-three')[0].innerHTML = quantityThree;
 
-        updateCartTotal ()
-    }else {
+        updateCartTotal()
+    } else {
         quantityThree = 1;
-        updateCartTotal ()
+        updateCartTotal()
     }
-    
 
-}) 
+
+})
 
 
 
@@ -153,6 +153,8 @@ var addToCartOne = document.getElementsByClassName('buy-btn-one')[0];
 
 addToCartOne.addEventListener('click', function () {
 
+    alert("Item added to the cart!")
+
     document.getElementById('cont-cart-item-one').className = 'cont-prod-cart';
 
     console.log("Has sumado un elemento a tu carrito");
@@ -162,7 +164,7 @@ addToCartOne.addEventListener('click', function () {
     console.log(quantityOne);
 
     document.getElementsByClassName('item-quantity-one')[0].innerHTML = quantityOne;
-    updateCartTotal ()
+    updateCartTotal()
 
 })
 
@@ -170,6 +172,8 @@ addToCartOne.addEventListener('click', function () {
 var addToCartTwo = document.getElementsByClassName('buy-btn-two')[0];
 
 addToCartTwo.addEventListener('click', function () {
+
+    alert("Item added to the cart!")
 
     document.getElementById('cont-cart-item-two').className = 'cont-prod-cart';
 
@@ -180,7 +184,7 @@ addToCartTwo.addEventListener('click', function () {
     console.log(quantityTwo);
 
     document.getElementsByClassName('item-quantity-two')[0].innerHTML = quantityTwo;
-    updateCartTotal ()
+    updateCartTotal()
 
 })
 
@@ -191,6 +195,8 @@ var addToCartThree = document.getElementsByClassName('buy-btn-three')[0];
 
 addToCartThree.addEventListener('click', function () {
 
+    alert("Item added to the cart!")
+
     document.getElementById('cont-cart-item-three').className = 'cont-prod-cart';
 
     console.log("Has sumado un elemento a tu carrito");
@@ -200,7 +206,7 @@ addToCartThree.addEventListener('click', function () {
     console.log(quantityThree);
 
     document.getElementsByClassName('item-quantity-three')[0].innerHTML = quantityThree;
-    updateCartTotal ()
+    updateCartTotal()
 
 })
 
@@ -213,7 +219,7 @@ addToCartThree.addEventListener('click', function () {
 
 // DELETE - 1
 var deleteOne = document.getElementsByClassName('del-one')[0];
-deleteOne.addEventListener('click', function (event){
+deleteOne.addEventListener('click', function (event) {
     deleteOne.parentElement.className = "delete";
     quantityOne = 0;
     document.getElementsByClassName('item-quantity-one').innerHTML = quantityOne;
@@ -223,33 +229,38 @@ deleteOne.addEventListener('click', function (event){
 
 // DELETE - 2
 var deleteTwo = document.getElementsByClassName('del-two')[0];
-deleteTwo.addEventListener('click', function (event){
+deleteTwo.addEventListener('click', function (event) {
     deleteTwo.parentElement.className = "delete";
     quantityTwo = 0;
-    updateCartTotal ()
+    updateCartTotal()
 })
 
 // DELETE - 3
 var deleteThree = document.getElementsByClassName('del-three')[0];
-deleteThree.addEventListener('click', function (event){
+deleteThree.addEventListener('click', function (event) {
     deleteThree.parentElement.className = "delete";
     quantityThree = 0;
-    updateCartTotal ()
+    updateCartTotal()
 })
 
 
 // PAY
 
 var payment = document.getElementById('pay');
-payment.addEventListener('click', function (event){
-    deleteThree.parentElement.className = "delete";
-    deleteTwo.parentElement.className = "delete";
-    deleteOne.parentElement.className = "delete";
-    quantityThree = 0;
-    quantityTwo = 0;
-    quantityOne = 0;
-    alert("Thanks for the payment!")
-    updateCartTotal ()
+payment.addEventListener('click', function (event) {
+    if (document.getElementById('total').innerHTML > 0) {
+        deleteThree.parentElement.className = "delete";
+        deleteTwo.parentElement.className = "delete";
+        deleteOne.parentElement.className = "delete";
+        quantityThree = 0;
+        quantityTwo = 0;
+        quantityOne = 0;
+        alert("Thanks for the payment!")
+        updateCartTotal()
+    } else {
+        alert("No items added  :(")
+    }
+
 })
 
 
@@ -257,11 +268,11 @@ payment.addEventListener('click', function (event){
 
 // Updating price
 
-updateCartTotal ()
+updateCartTotal()
 
 // Cart total calculation
-function updateCartTotal () {
-    
+function updateCartTotal() {
+
     var total = 0;
 
     total = total + (quantityOne * priceOne) + (quantityTwo * priceTwo) + (quantityThree * priceThree);
